+++ src/main/java/com/arangodb/springframework/annotation/EnableArangoAuditing.java
@@ -34,6 +34,8 @@ import org.springframework.data.domain.AuditorAware;
 import com.arangodb.springframework.config.ArangoAuditingRegistrar;
 
 /**
+ * Annotation to activate auditing in ArangoDB via annotation configuration.
+ *
  * @author Mark Vollmary
  *
  */
@@ -47,30 +49,22 @@ public @interface EnableArangoAuditing {
 
 	/**
 	 * Configures the {@link AuditorAware} bean to be used to lookup the current principal.
-	 *
-	 * @return
 	 */
 	String auditorAwareRef() default "";
 
 	/**
 	 * Configures whether the creation and modification dates are set. Defaults to {@literal true}.
-	 *
-	 * @return
 	 */
 	boolean setDates() default true;
 
 	/**
 	 * Configures whether the entity shall be marked as modified on creation. Defaults to {@literal true}.
-	 *
-	 * @return
 	 */
 	boolean modifyOnCreate() default true;
 
 	/**
 	 * Configures a {@link DateTimeProvider} bean name that allows customizing the {@link org.joda.time.DateTime} to be
 	 * used for setting creation and modification dates.
-	 *
-	 * @return
 	 */
 	String dateTimeProviderRef() default "";