+++ src/main/java/com/arangodb/springframework/annotation/PersistentIndex.java
@@ -27,6 +27,9 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 /**
+ * Annotation to define given fields to be indexed using ArangoDB's Persistent
+ * index.
+ *
  * @author Mark Vollmary
  *
  */
@@ -36,17 +39,17 @@ import java.lang.annotation.Target;
 public @interface PersistentIndex {
 
 	/**
-	 * @return A list of attribute paths
+	 * A list of attribute paths
 	 */
 	String[] fields();
 
 	/**
-	 * @return if true, then create a unique index
+	 * If {@literal true}, then create a unique index
 	 */
 	boolean unique() default false;
 
 	/**
-	 * @return if true, then create a sparse index
+	 * If {@literal true}, then create a sparse index
 	 */
 	boolean sparse() default false;